+++ checkpoint/orbax/checkpoint/experimental/v1/_src/context/context.py
@@ -101,6 +101,7 @@ class Context(epy.ContextManager):
     checkpointables_options: Options for controlling checkpointables behavior.
     pathways_options: Options for Pathways checkpointing.
     checkpoint_layout: The layout of the checkpoint. Defaults to ORBAX.
+    deletion_options: Options for controlling deletion behavior.
   """
 
   def __init__(
@@ -115,6 +116,7 @@ class Context(epy.ContextManager):
       checkpointables_options: options_lib.CheckpointablesOptions | None = None,
       pathways_options: options_lib.PathwaysOptions | None = None,
       checkpoint_layout: options_lib.CheckpointLayout | None = None,
+      deletion_options: options_lib.DeletionOptions | None = None,
   ):
     self._pytree_options = pytree_options or (
         context.pytree_options if context else options_lib.PyTreeOptions()
@@ -145,6 +147,9 @@ class Context(epy.ContextManager):
         context.checkpoint_layout
         if context
         else options_lib.CheckpointLayout.ORBAX
+    )
+    self._deletion_options = deletion_options or (
+        context.deletion_options if context else options_lib.DeletionOptions()
     )
 
   @property
@@ -178,6 +183,10 @@ class Context(epy.ContextManager):
   @property
   def checkpoint_layout(self) -> options_lib.CheckpointLayout:
     return self._checkpoint_layout
+
+  @property
+  def deletion_options(self) -> options_lib.DeletionOptions:
+    return self._deletion_options
 
   def operation_id(self) -> str:
     return synchronization.OperationIdGenerator.get_current_operation_id()