+++ checkpoint/orbax/checkpoint/experimental/v1/_src/context/options.py
@@ -494,6 +494,44 @@ class PathwaysOptions:
   checkpointing_impl: pathways_types.CheckpointingImpl | None = None
 
 
+@dataclasses.dataclass(frozen=True, kw_only=True)
+class DeletionOptions:
+  """Options used to configure checkpoint deletion behavior.
+
+  Attributes:
+    gcs_deletion_options: Deletion options specific to GCS.
+  """
+
+  @dataclasses.dataclass(frozen=True, kw_only=True)
+  class GcsDeletionOptions:
+    """Deletion options specific to GCS.
+
+    Attributes:
+      todelete_full_path: A path relative to the bucket root for "soft-deleting"
+        checkpoints on Google Cloud Storage (GCS). Instead of being permanently
+        removed, checkpoints are moved to this new location within the same
+        bucket. This is useful if direct deletion on GCS is time-consuming, as
+        it allows an external component to manage the actual removal.
+
+        This option gathers all "deleted" items in a centralized path at the
+        bucket level for future cleanup.
+
+        For instance, if a checkpoint is in
+        gs://my-bucket/experiments/run1/, providing the value 'trash' will move
+        a deleted step to gs://my-bucket/trash/<step_id>. Useful when direct
+        deletion is time consuming. It gathers all deleted items in a
+        centralized path for future cleanup.
+    """
+
+    todelete_full_path: str | None = None
+
+
+  gcs_deletion_options: GcsDeletionOptions = dataclasses.field(
+      default_factory=GcsDeletionOptions
+  )
+
+
+
 class CheckpointLayout(enum.Enum):
   """The layout of the checkpoint.